Add CommandTypes.
Now it is possible to add Command classes as Server commands.
So server commands don't have to be simple callback functions
and a lot of code can be reused, since Execute is called with
null when it is a server command.

Furthermore adjust the existing commands to work with this new
paradigm.